refactor(lotto): 테스트 클래스 패키지 이동
