+++ src/test/kotlin/lotto/domain/lotto/LottoNumberTest.kt
@@ -1,6 +1,5 @@
-package lotto.domain
+package lotto.domain.lotto
 
-import lotto.domain.lotto.LottoNumber
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.params.ParameterizedTest
